Develop

+++ Compiler/include/Symbols.h
@@ -9,11 +9,10 @@ class Symbols {
 public:
 	Symbols(Symbols* parent = nullptr);
 
-	typedef std::unordered_map<std::string, std::variant<float, int, bool, Function*>> SymbolsMap;
+	typedef std::unordered_map<std::string, std::variant<double, int, bool, Function*>> SymbolsMap;
 
 	SymbolsMap::iterator get(const std::string& name);
-	void set(const std::string& name, std::variant<float, int, bool, Function*> value);
-
+	void set(const std::string& name, std::variant<double, int, bool, Function*> value);
 	bool remove(const std::string& name);
 	
 	SymbolsMap symbols;
